Refactor: Moving `hd_active` to its own package.
