+++ hd_active/hd_active/hd_active_state.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class HdActionState(str, Enum):
+    Start = 'Start'
+    Stop = 'Stop'
